+++ com.julienpoeschl.settings/Runtime/Setting.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace Settings
+{
+    public class Setting : ScriptableObject { }
+}